Extract the Facebook oauth strategy patch to its own file
Could have been moved to an entirely different strategy that inherited
from the FB one, but that would have been a breaking change. In this
way we at least improve the relation between filenames and defined
constants.
